[test] SampleData: move to a dedicated package
